backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
firebase-admin==6.5.0
python-multipart==0.0.9
email-validator==2.1.1
aiofiles==23.2.1
